pyproject.toml
[project]
name = "postgresql-mcp-python"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14.2"
dependencies = [
    "asyncpg>=0.31.0",
    "mcp[cli]>=1.26.0",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
]

[dependency-groups]
dev = [
    "ruff>=0.14.14",
]
